Refactor code and add some docstrings.
