+++ sdks/python/apache_beam/ml/anomaly/base.py
@@ -37,6 +37,10 @@ __all__ = [
     "EnsembleAnomalyDetector"
 ]
 
+DEFAULT_NORMAL_LABEL = 0
+DEFAULT_OUTLIER_LABEL = 1
+DEFAULT_MISSING_LABEL = -2
+
 
 @dataclass(frozen=True)
 class AnomalyPrediction():
@@ -79,9 +83,9 @@ class ThresholdFn(abc.ABC):
   """
   def __init__(
       self,
-      normal_label: int = 0,
-      outlier_label: int = 1,
-      missing_label: int = -2):
+      normal_label: int = DEFAULT_NORMAL_LABEL,
+      outlier_label: int = DEFAULT_OUTLIER_LABEL,
+      missing_label: int = DEFAULT_MISSING_LABEL):
     self._normal_label = normal_label
     self._outlier_label = outlier_label
     self._missing_label = missing_label
